+++ src/addrman.h
@@ -24,78 +24,6 @@ class AddrManImpl;
 /** Default for -checkaddrman */
 static constexpr int32_t DEFAULT_ADDRMAN_CONSISTENCY_CHECKS{0};
 
-/**
- * Extended statistics about a CAddress
- */
-class CAddrInfo : public CAddress
-{
-public:
-    //! last try whatsoever by us (memory only)
-    int64_t nLastTry{0};
-
-    //! last counted attempt (memory only)
-    int64_t nLastCountAttempt{0};
-
-private:
-    //! where knowledge about this address first came from
-    CNetAddr source;
-
-    //! last successful connection by us
-    int64_t nLastSuccess{0};
-
-    //! connection attempts since last successful attempt
-    int nAttempts{0};
-
-    //! reference count in new sets (memory only)
-    int nRefCount{0};
-
-    //! in tried set? (memory only)
-    bool fInTried{false};
-
-    //! position in vRandom
-    mutable int nRandomPos{-1};
-
-    friend class AddrManImpl;
-    friend class CAddrManDeterministic;
-
-public:
-
-    SERIALIZE_METHODS(CAddrInfo, obj)
-    {
-        READWRITEAS(CAddress, obj);
-        READWRITE(obj.source, obj.nLastSuccess, obj.nAttempts);
-    }
-
-    CAddrInfo(const CAddress &addrIn, const CNetAddr &addrSource) : CAddress(addrIn), source(addrSource)
-    {
-    }
-
-    CAddrInfo() : CAddress(), source()
-    {
-    }
-
-    //! Calculate in which "tried" bucket this entry belongs
-    int GetTriedBucket(const uint256 &nKey, const std::vector<bool> &asmap) const;
-
-    //! Calculate in which "new" bucket this entry belongs, given a certain source
-    int GetNewBucket(const uint256 &nKey, const CNetAddr& src, const std::vector<bool> &asmap) const;
-
-    //! Calculate in which "new" bucket this entry belongs, using its default source
-    int GetNewBucket(const uint256 &nKey, const std::vector<bool> &asmap) const
-    {
-        return GetNewBucket(nKey, source, asmap);
-    }
-
-    //! Calculate in which position of a bucket to store this entry.
-    int GetBucketPosition(const uint256 &nKey, bool fNew, int nBucket) const;
-
-    //! Determine whether the statistics about this entry are bad enough so that it can just be deleted
-    bool IsTerrible(int64_t nNow = GetAdjustedTime()) const;
-
-    //! Calculate the relative chance this entry should be given when selecting nodes to connect to
-    double GetChance(int64_t nNow = GetAdjustedTime()) const;
-};
-
 /** Stochastic address manager
  *
  * Design goals:
